move startup tracing into the SDK
