+++ src/Sentry.Unity/SentryUnity.cs
@@ -1,6 +1,7 @@
 using System;
 using System.ComponentModel;
 using Sentry.Extensibility;
+using UnityEngine;
 
 namespace Sentry.Unity;
 
@@ -10,6 +11,12 @@ namespace Sentry.Unity;
 public static class SentryUnity
 {
     private static SentryUnitySdk? UnitySdk;
+
+    [RuntimeInitializeOnLoadMethod(RuntimeInitializeLoadType.AfterSceneLoad)]
+    private static void DoTheThing()
+    {
+        Debug.Log("What the actual fuck.");
+    }
 
     /// <summary>
     /// Initializes Sentry Unity SDK while configuring the options.
